@@ -20,9 +20,9 @@ module Spree
 
     scope :from_credit_card, lambda { where(:source_type => 'Spree::CreditCard') }
     scope :with_state, lambda { |s| where(:state => s) }
-    scope :completed, with_state('completed')
-    scope :pending, with_state('pending')
-    scope :failed, with_state('failed')
+    scope :completed, -> { with_state('completed') }
+    scope :pending, -> { with_state('pending') }
+    scope :failed, -> { with_state('failed') }
 
     # order state machine (see http://github.com/pluginaweek/state_machine/tree/master for details)
     state_machine :initial => 'checkout' do
